Extract sorting change listener to separate class
